requirements.txt
numpy==1.14.2
matplotlib==2.2.0
tensorflow-gpu==1.9.0
tqdm==4.23.4
pyyaml==3.13
nose==1.3.7
